08_FitStep_API/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai>=2.0.0
chromadb==1.5.8
python-dotenv==1.0.1
pydantic>=2.0.0
mysql-connector-python==8.3.0
